Represent review and violation per domain
This resource serve review and violation data aggregated per
domain.

To make things a bit faster, a new field `review_is_active' was
added to the violation model and a new index was also created in
the respective table. See the included migration for detail.

Just before setting a Page's last review to inactive the worker
also sets all it's violations to inactive.